requirements.txt
Django==5.2.7
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
psycopg2-binary==2.9.11
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
django_extensions==4.1
django-filter==25.2
pillow==12.0.0
requests==2.32.5
django-jazzmin==3.0.1